requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.13.0
scikit-learn>=1.2.2
lightning>=2.0.0
torchmetrics>=0.11.0
hydra-core>=1.3.2
omegaconf>=2.3.0
pandas>=2.0.0
pyarrow>=12.0.0
tensorboard>=2.13.0
tensorboardX>=2.6.0
rich>=13.0.0
opencv-python-headless>=4.8.0 